Refactor: rename events package, rename listeners
