+++ src/main/java/com/sparkedhost/accuratereadings/Main.java
@@ -2,7 +2,7 @@ package com.sparkedhost.accuratereadings;
 
 import com.sparkedhost.accuratereadings.commands.*;
 import com.sparkedhost.accuratereadings.config.Settings;
-import com.sparkedhost.accuratereadings.events.RestartCmdEvent;
+import com.sparkedhost.accuratereadings.listeners.RestartCommandListener;
 import com.sparkedhost.accuratereadings.managers.PlaceholderAPIManager;
 import com.sparkedhost.accuratereadings.managers.PterodactylManager;
 import com.sparkedhost.accuratereadings.managers.TaskManager;
@@ -77,7 +77,7 @@ public class Main extends JavaPlugin {
 
         if (getConfig().getBoolean("enableRestartCmd")) {
             // TODO Switch back to onCommand
-            getServer().getPluginManager().registerEvents(new RestartCmdEvent(), this);
+            getServer().getPluginManager().registerEvents(new RestartCommandListener(), this);
         }
 
         log(Level.INFO, "Loaded all the commands. Connecting to the panel using '" + panelUrl + "'...");
