+++ src/main/java/com/sparkedhost/accuratereadings/listeners/RestartCommandListener.java
@@ -1,4 +1,4 @@
-package com.sparkedhost.accuratereadings.events;
+package com.sparkedhost.accuratereadings.listeners;
 
 import com.sparkedhost.accuratereadings.Main;
 import org.bukkit.event.EventHandler;
@@ -6,7 +6,7 @@ import org.bukkit.event.EventPriority;
 import org.bukkit.event.Listener;
 import org.bukkit.event.player.PlayerCommandPreprocessEvent;
 
-public class RestartCmdEvent implements Listener {
+public class RestartCommandListener implements Listener {
     @EventHandler(priority = EventPriority.HIGHEST)
     public void onRestartCommand(PlayerCommandPreprocessEvent e) {
         if (e.getMessage().equals("/restart") && e.getPlayer().hasPermission("readings.restart")) {
